rag_minimal/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
python-dotenv>=1.0.0
pydantic>=2.5.0

# Vector store and embeddings
chromadb>=0.4.18
sentence-transformers>=2.2.2

# Data processing
pandas>=2.1.0
numpy>=1.24.0
beautifulsoup4>=4.12.0
python-docx>=0.8.11
PyPDF2>=3.0.0

# Text processing
nltk>=3.8.1
spacy>=3.7.0

# API and web
fastapi>=0.104.0
uvicorn>=0.24.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0 